% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.R
\name{as.data.table.scalib}
\alias{as.data.table.scalib}
\title{Coerce to \code{data.table}}
\usage{
\method{as.data.table}{scalib}(x, ...)
}
\arguments{
\item{x}{An object of class \code{scalib} (see \link{scalib}).}

\item{...}{currently not used}
}
\value{
a \link[data.table:data.table]{data.table} object
}
\description{
Convert \code{scalib} object into a flat data.table containing
input and output data.
}
