% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_slope_hare.R
\name{as_tibble.survival_calib_slope}
\alias{as_tibble.survival_calib_slope}
\title{Coerce to a Tibble}
\usage{
as_tibble.survival_calib_slope(x, ...)
}
\arguments{
\item{x}{an object of class \code{survival_calib_slope}}

\item{...}{not used}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Coerce to a Tibble
}
\examples{

}
