% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_test_gnd_auto.R
\name{calib_test_gnd}
\alias{calib_test_gnd}
\title{Greenwood-Nam-D'Agostino (GND) Test on Autopilot}
\usage{
calib_test_gnd(
  predicted_risk,
  event_time,
  event_status,
  time_predict,
  group_method = "lump",
  group_count_init = 10,
  group_count_min = 2,
  group_min_events_warn = 5,
  group_min_events_stop = 2,
  verbose = 0
)
}
\arguments{
\item{predicted_risk}{(\emph{numeric vector}) predicted risk values for the
event of interest at or before a given time.}

\item{event_time}{(\emph{numeric vector}) observed event times}

\item{event_status}{(\emph{numeric vector}) observed event status. The values
of this vector should be 0 (event censored) and 1 (event observed).}

\item{time_predict}{(\emph{numeric value}) the time of risk prediction.}

\item{group_method}{(\emph{character value}; 'lump' or 'redo') If 'lump', then
a 'lumping' procedure will be applied whenever a group has less than
\code{group_min_events_warn} events. The lumping procedure will identify
whichever group has the lowest event count and assign members of that
group to the group with too few events. If 'redo', then the groups
will be re-done using \code{cut_percentile} but with one less group.
'lump' is the default as this method was studied by Demler et al.}

\item{group_count_init}{(\emph{integer value}) the initial number of
groups to form based on percentiles of \code{predicted_risk} values.}

\item{group_count_min}{(\emph{integer value}) the minimum number of
groups to attempt running the GND test with. If}

\item{group_min_events_warn}{(\emph{numeric value}) The lowest event count
within a risk group that will not cause a warning (see details).}

\item{group_min_events_stop}{(\emph{numeric value}) The lowest event count
within a risk group that will not cause a hard stop (see details).}

\item{verbose}{(\emph{integer value}) If 0, no output will be printed.
If 1, then output from group reduction will be printed.
If 2, then output from group reduction \strong{and} \link{calib_test_gnd_manual}
will be printed.}
}
\value{
an object of class 'survival_calib_test'
}
\description{
The GND test is based on groupwise error rates for observed versus
expected risk, and the test is unstable when any group contains
< 5 events (see Demler et al.). The purpose of \code{calib_test_gnd} is
to automatically check event counts in risk groups and, if
necessary, collapse risk groups so that every group has an
event count greater than a given threshold.
}
\examples{
# packages
library(riskRegression)
library(survival)

# set index for training data
train_index <- 1:3000

# fit cox PH model using train data
risk_mdl <- coxph(data = flchain[train_index, ],
                  x = TRUE,
                  formula = Surv(futime, death)~age+sex+flc.grp+lambda)

# set time of prediction
time_predict <- 4000

# predict risk in the testing data
risk_pred <- predictRisk(risk_mdl,
                         newdata = flchain[-train_index, ],
                         times = time_predict)

# run GND test using 10 groups; apply some rules:
#  1. require at least 40 events per group
#  2. try to create 10 groups, and lump the lowest event frequency groups
#     if any groups have less than 40 events.
#  3. Hard stop if this is not obtainable with 5 or more groups.
#  4. set verbose = 2 to see every detail.
# (note that these rules are not a recommendation on how to use the test;
#  this example just shows the mechanics of the automatic group reduction.
#  The recommended values are the default values.)

calib_test_gnd(predicted_risk = risk_pred,
              event_time = flchain$futime[-train_index],
              event_status = flchain$death[-train_index],
              time_predict = time_predict,
              group_min_events_warn = 40,
              group_count_init = 10,
              group_count_min = 5,
              verbose = 2)
}
\references{
Demler, O.V., Paynter, N.P. and Cook, N.R., 2015. Tests of calibration and
goodness‐of‐fit in the survival setting. \emph{Statistics in medicine}, 34(10),
pp.1659-1680. DOI: 10.1002/sim.6428
}
