% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_test_gnd_manual.R
\name{calib_test_gnd_manual}
\alias{calib_test_gnd_manual}
\title{Greenwood-Nam-D'Agostino (GND) Test}
\usage{
calib_test_gnd_manual(
  predicted_risk,
  event_status,
  event_time,
  time_predict,
  group,
  group_min_events_warn = 5,
  group_min_events_stop = 2,
  verbose = 0,
  id_value = NULL
)
}
\arguments{
\item{predicted_risk}{(\emph{numeric vector}) predicted risk values for the
event of interest at or before a given time.}

\item{event_status}{(\emph{numeric vector}) observed event status. The values
of this vector should be 0 (event censored) and 1 (event observed).}

\item{event_time}{(\emph{numeric vector}) observed event times}

\item{time_predict}{(\emph{numeric value}) the time of risk prediction.}

\item{group}{(\emph{numeric vector}) risk group assignment. The GND chi-square
test statistic is based on the difference between predicted and observed
event counts in each risk group.}

\item{group_min_events_warn}{(\emph{numeric value}) The lowest event count
within a risk group that will not cause a warning (see details).}

\item{group_min_events_stop}{(\emph{numeric value}) The lowest event count
within a risk group that will not cause a hard stop (see details).}

\item{verbose}{(\emph{integer value}) If 0, no output will be printed.
If 1, some details will be printed. If 2, all details will be printed.}
}
\value{
an object of class 'survival_calib_test'
}
\description{
Demler et al developed a modification of the Nam-D'Agostino test that has
nominal type 1 error rates with censored outcomes. The GND test is also
applicable in settings where the proportional hazards assumption is invalid.
}
\details{
\strong{Minimum event counts for risk groups}: Low event counts within any risk
group may cause high variability in the GND test results. It is recommended
that all risk groups have at least 5 events, and this is why the default
value of \code{group_min_events_warn} is 5. If there are less than 2 events
in any group, the GND test is unstable and risk groups should be collapsed.
Therefore, the default value of \code{group_min_events_stop} is 2.
}
\examples{

library(survival)
library(riskRegression)
train_index <- 1:3000

risk_mdl <- coxph(data = flchain[train_index, ],
                  x = TRUE,
                  formula = Surv(futime, death)~age+sex+flc.grp+lambda)

time_predict <- 4000

risk_pred <- predictRisk(risk_mdl,
                         newdata = flchain[-train_index, ],
                         times = time_predict)

#split into deciles
risk_groups <- predrisk_grp_prcnt(risk_pred, g = 10)

calib_test_gnd_manual(predicted_risk = risk_pred,
                event_time = flchain$futime[-train_index],
                event_status = flchain$death[-train_index],
                time_predict = time_predict,
                group = risk_groups,
                verbose = 0)
}
\references{
Demler, O.V., Paynter, N.P. and Cook, N.R., 2015. Tests of calibration and
goodness‐of‐fit in the survival setting. \emph{Statistics in medicine}, 34(10),
pp.1659-1680. DOI: 10.1002/sim.6428
}
