% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnd_test_auto.R
\name{gnd_test_auto}
\alias{gnd_test_auto}
\title{Modified Greenwood-Nam D'Agostino Test}
\usage{
gnd_test_auto(
  predicted_risk,
  event_time,
  event_status,
  time_predict,
  group_count_init = 10,
  group_count_min = 2,
  group_min_events_warn = 5,
  group_min_events_stop = 2,
  verbose = 1
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Modified Greenwood-Nam D'Agostino Test
}
