% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predrisk_grp_prcnt.R
\name{predrisk_grp_prcnt}
\alias{predrisk_grp_prcnt}
\title{Cut data into percentile groups}
\usage{
predrisk_grp_prcnt(x, g, na.rm = FALSE, type = 7)
}
\arguments{
\item{x}{(\emph{numeric vector}) a numeric vector which will be converted to
a vector of integers by cutting.}

\item{g}{(\emph{numeric value}) the desired number of percentile groups. For
example, \code{g = 10} gives decile groups.}

\item{na.rm}{(\emph{logical value}) if \code{TRUE}, any \code{NA} and \code{NaN} values
are removed from \code{x} before the quantiles are computed.}

\item{type}{(\emph{integer value between 1 and 9}) see \link[stats:quantile]{quantile} for
details.}
}
\value{
a numeric vector with integer values.
}
\description{
\code{predrisk_grp_prcnt()} is a simple wrapper for \code{cut} that is meant to be used
to create an input value for \code{\link[=calib_test_gnd_manual]{calib_test_gnd_manual()}}  (specifically, the
\code{group} input argument).
}
\examples{

predrisk_grp_prcnt(sort(rnorm(100)), g = 10)

}
