% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_test_print.R
\name{print.survival_calib_test}
\alias{print.survival_calib_test}
\title{Print Calibration Test Results}
\usage{
\method{print}{survival_calib_test}(x, ...)
}
\arguments{
\item{x}{an object of class \code{survival_calib_test}}

\item{...}{additional arguments. Eligible inputs are:
\itemize{
\item \strong{digits}: minimal number of significant digits to print.
\item \strong{max_rows_to_print}: maximum number of rows to print
}}
}
\value{
nothing. Output is printed to console.
}
\description{
Print Calibration Test Results
}
