% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalib_gnd.R, R/vec_gnd.R
\name{scalib_gnd}
\alias{scalib_gnd}
\alias{scalib_gnd_manual}
\alias{vec_gnd}
\alias{vec_gnd_manual}
\title{Greenwood-Nam-D'Agostino (GND) Test}
\usage{
scalib_gnd(
  scalib_object,
  group_method = "lump",
  group_count_init = 10,
  group_count_min = 2,
  group_min_events = 5,
  verbose = 0
)

scalib_gnd_manual(
  scalib_object,
  pred_risk_col,
  group,
  group_min_events_warn = 5,
  group_min_events_stop = 2,
  verbose = 0
)

vec_gnd(
  pred_risk,
  pred_horizon,
  event_status,
  event_time,
  group_method = "lump",
  group_count_init = 10,
  group_count_min = 2,
  group_min_events = 5,
  verbose = 0
)

vec_gnd_manual(
  pred_risk,
  pred_horizon,
  event_status,
  event_time,
  group,
  group_min_events_warn = 5,
  group_min_events_stop = 2,
  verbose = 0
)
}
\arguments{
\item{scalib_object}{An object of class \code{scalib}
(see \link{scalib}).}

\item{group_method}{(\emph{character value}; 'lump' or 'redo') If 'lump', then
a 'lumping' procedure will be applied whenever a group has less than
\code{group_min_events_warn} events. The lumping procedure will identify
whichever group has the lowest event count and assign members of that
group to the group with too few events. If 'redo', then the groups
will be re-done using \link{predrisk_grp_prcnt} but with one less group.
'lump' is the default as this method was studied by Demler et al.}

\item{group_count_init}{(\emph{integer value}) the initial number of
groups to form based on percentiles of \code{predicted_risk} values.}

\item{group_count_min}{(\emph{integer value}) the minimum number of
groups to attempt running the GND test with.}

\item{verbose}{(\emph{integer value}) If 0, no output will be printed.
If 1, some output will be printed. If 2, \emph{all} output will be printed.}

\item{pred_risk_col}{(\emph{character value}) the column name of the variable
that contains predicted risk values. This variable should be in
\code{scalib_object$data_inputs}.}

\item{group_min_events_warn}{(\emph{numeric value}) The lowest event count
within a risk group that will not cause a warning (see details).}

\item{group_min_events_stop}{(\emph{numeric value}) The lowest event count
within a risk group that will not cause a hard stop (see details).}

\item{pred_risk}{(\emph{numeric vector}, \emph{list}, \emph{data frame}, or \emph{matrix})
predicted risk values for the event at or before \code{pred_horizon}}

\item{pred_horizon}{(\emph{numeric value}) the time of risk prediction.}

\item{event_status}{(\emph{numeric vector}) observed event status. The values
of this vector should be 0 (event censored) and 1 (event observed).}

\item{event_time}{(\emph{numeric vector}) observed event times}

\item{group_col}{(\emph{character value}) the column name of the variable
that contains risk group values. This variable should be in
\code{scalib_object$data_inputs}}
}
\value{
an object of class \code{scalib} (see \link{scalib})
}
\description{
Demler et al developed a modification of the Nam-D'Agostino test that has
nominal type 1 error rates with censored outcomes. The
Greenwood-Nam-D'Agostino (GND) test is applicable in settings where
the proportional hazards assumption is invalid. The GND test is based
on groupwise error rates for observed versus expected risk, and the
test is unstable when any group contains < 5 events (see Demler et al.).
}
\details{
\code{scalib_gnd} automatically forms risk groups, checks event counts
in risk groups and, if necessary, collapses risk groups so that
every group has an event count greater than a given threshold.

\code{scalib_gnd_manual} completes the GND test, but requires the user to
\enumerate{
\item create a column in the scalib_object$data_inputs that contains
risk group labels (see \link{predrisk_grp_prcnt})
\item specify the column name of the risk group column and the
predicted risk column.
}

\strong{Minimum event counts for risk groups}: Low event counts within any risk
group may cause high variability in the GND test results. It is recommended
that all risk groups have at least 5 events, and this is why the default
value of \code{group_min_events_warn} is 5. If there are less than 2 events
in any group, the GND test is unstable and risk groups should be collapsed.
Therefore, the default value of \code{group_min_events_stop} is 2.
}
\examples{
# packages
library(riskRegression)
library(survival)

# set index for training data
train_index <- 1:3000

# fit cox PH model using train data
risk_mdl <- coxph(data = flchain[train_index, ],
                  x = TRUE,
                  formula = Surv(futime, death)~age+sex+flc.grp+lambda)

# set time of prediction
pred_horizon <- 4000

# predict risk in the testing data
pred_risk <- predictRisk(risk_mdl,
                         newdata = flchain[-train_index, ],
                         times = pred_horizon)

# create a scalib object



# run GND test using 10 groups; apply some rules:
#  1. require at least 40 events per group
#  2. try to create 10 groups, and lump the lowest event frequency groups
#     if any groups have less than 40 events.
#  3. Hard stop if this is not obtainable with 5 or more groups.
#  4. set verbose = 2 to see every detail.
# (note that these rules are not a recommendation on how to use the test;
#  this example just shows the mechanics of the automatic group reduction.
#  The recommended values are the default values.)

scalib_gnd(predicted_risk = pred_risk,
              event_time = flchain$futime[-train_index],
              event_status = flchain$death[-train_index],
              pred_horizon = pred_horizon,
              group_min_events_warn = 40,
              group_count_init = 10,
              group_count_min = 5,
              verbose = 2)
}
\references{
Demler, O.V., Paynter, N.P. and Cook, N.R., 2015. Tests of calibration and
goodness‐of‐fit in the survival setting. \emph{Statistics in medicine}, 34(10),
pp.1659-1680. DOI: 10.1002/sim.6428
}
