% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.calib-package.R
\docType{package}
\name{survival.calib-package}
\alias{survival.calib}
\alias{survival.calib-package}
\title{survival.calib: Tests of Calibration in the Survival Setting}
\description{
There are multiple tests for mis-calibration in the setting of binary outcomes, but these tests may not maintain a nominal type 1 error rate in the presence of censored outcomes. This package offers tests for calibration that maintain nominal type 1 error rate in the presence of censoring and do not require the assumption of proportional hazards.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bcjaeger.github.io/survival.calib}
  \item \url{https://github.com/bcjaeger/survival.calib}
  \item Report bugs at \url{https://github.com/bcjaeger/survival.calib/issues}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

Authors:
\itemize{
  \item Olga Demler \email{odemler@bwh.harvard.edu} (\href{https://orcid.org/0000-0003-3355-3210}{ORCID})
}

}
\keyword{internal}
