% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.R
\name{as.data.table.scalib}
\alias{as.data.table.scalib}
\title{Coerce to \code{data.table}}
\usage{
\method{as.data.table}{scalib}(x, ...)
}
\arguments{
\item{x}{An object of class \code{scalib} (see \link{scalib}).}

\item{...}{currently not used}
}
\value{
a \link[data.table:data.table]{data.table} object
}
\description{
Convert \code{scalib} object into a flat data.table containing
input and output data.
}
\examples{


 library(data.table)

 sc <- scalib(pred_risk = pbc_scalib$predrisk,
              pred_horizon = 2500,
              event_time = pbc_scalib$test$time,
              event_status = pbc_scalib$test$status)

 print(sc)

 as.data.table(sc)

 sc_gnd <- scalib_gnd(sc)

 as.data.table(sc_gnd)


}
