% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnd_test_manual.R
\name{print.survival.calib_gnd_test}
\alias{print.survival.calib_gnd_test}
\title{Print GND Test Results}
\usage{
\method{print}{survival.calib_gnd_test}(x, digits = 3)
}
\arguments{
\item{x}{an object of class \code{survival.calib_gnd_test}}

\item{digits}{minimal number of significant digits to print.}
}
\value{
nothing. Output is printed to console.
}
\description{
Print GND Test Results
}
