% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_calib_gnd_test.R
\name{print.survival.calib_gnd_test}
\alias{print.survival.calib_gnd_test}
\title{Print GND Test Results}
\usage{
\method{print}{survival.calib_gnd_test}(x, ...)
}
\arguments{
\item{x}{an object of class \code{survival.calib_gnd_test}}

\item{...}{additional arguments. Eligible inputs are:
\itemize{
\item \strong{digits}: minimal number of significant digits to print.
\item \strong{max_rows_to_print}: maximum number of rows to print
}}
}
\value{
nothing. Output is printed to console.
}
\description{
Print GND Test Results
}
