% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalib.R
\name{scalib}
\alias{scalib}
\title{Survival calibration objects}
\usage{
scalib(pred_risk, pred_horizon, event_status, event_time)
}
\arguments{
\item{pred_risk}{(\emph{numeric vector}, \emph{list}, \emph{data frame}, or \emph{matrix})
predicted risk values for the event at or before \code{pred_horizon}}

\item{pred_horizon}{(\emph{numeric value}) the time of risk prediction.}

\item{event_status}{(\emph{numeric vector}) observed event status. The values
of this vector should be 0 (event censored) and 1 (event observed).}

\item{event_time}{(\emph{numeric vector}) observed event times}
}
\value{
an object of class 'scalib'
}
\description{
Survival calibration objects
}
